+++ src/index.node.ts
@@ -1,0 +1,6 @@
+import { randomFillSync } from "node:crypto";
+import { setBytesProvider } from "./internal/random.js";
+
+setBytesProvider((length) => randomFillSync(new Uint8Array(length)));
+
+export * from "./index.js";