+++ tsup.config.ts
@@ -1,12 +1,12 @@
 import { defineConfig } from "tsup";
 
 export default defineConfig({
-  entry: ["src/index.ts"],
+  entry: ["src/index.ts", "src/index.node.ts"],
   format: ["esm", "cjs"],
   dts: true,
   clean: true,
   minify: false,
   sourcemap: true,
   treeshake: true,
-  target: "es2021",
+  target: "es2020",
 });